수정 - 패킷 구조 개선 헤더 부분 255까지 데이터는 255이후   ( 255바이트정도는 큰 데이터가 아니기때문에 추가 함     추 후에 기능 구현 고려 )

+++ WFAAgent.Framework/DataType.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using WFAAgent.Framework.Net.Sockets;
 
 namespace WFAAgent.Framework
 {
@@ -10,7 +11,7 @@ namespace WFAAgent.Framework
     {
         public static readonly DataType UserData = new DataType(DataContext.UserData);
         private int _Type;
-        private DataType(int type)
+        private DataType(short type)
         {
             DataContext.CheckedType(type);
             this._Type = type;
